Adding MatConstIteratorBase as a base class for MatConstIterator
This class has common members for the classes MatConstIterator and _MatConstIterator that will soon not be in a linear inheritance order anymore.
The idea is to have a base class with the common members, because elementSize won't be needed for the templated version.

Additionally member pointers are initialized with nullptr instead of 0, and some constructors are defaulted