Adding template meta programming backend and a header for forwarding the STL and using it (efficiently!) with openCV iterators
The idea behind it is two fold:
Because we can't know at compile time if our matrix (described by the iterator) is contiguous or not, a function cv::detail::__iterators__replaceable takes a variadic template pack and looks if there are opencv iterators. If yes it checks if they are contiguous.
Only if all present iterators point to a contiguous matrix, we get true. For everything else (including no opencv iterators in the pack) we get false.
This is tested with unit tests in the CORE_stl_forward suite

The second part consists of making a tuple with every iterator replaced by its (casted) pointer.
This part is limited to up to 10 input arguments, as I couldn't make it completely generic due to limitations on decltype(auto) in C++11. Starting from C++14 it is easy to replace however in a completely generic manner.

We then forward either the tuple or the original arguments to the stl function. Like this we get good performance (i.e. pointers) for where its possible, while staying safe with the iterators for all the cases, where this is not possible, combining the best of the two worlds.

Tests for this are added.

A new folder stl can now be found in core, containing the forwarded algorithms. This was changed in the CMakeLists.txt. The utility functions are added to the detail headers in core.