Adding find function as a prototype for stl functions returning iterators.
Here it is a bit more complicated, as we would return a pointer normally. So we differentiate as to when we need to return an opencv iterator by actually letting the algorithm run with pointers and then adding the difference to the begin pointer to the begin iterator, so that from the outside it's not noticeable that the algo runs with pointers instead of iterators.
This requires another constexpr function __get_first_cv_it_index for which tests are added.